Optimize the priority of ServiceManager(2/2).
During the application startup, we find that the Servicemanager thread is frequently preempted
due to its low priority, resulting in slow application startup and responsiveness.
We hope to optimize these issues by adjusting the priority to RT(98) of the SM
to avoid time-consuming binder in clients to ServiceManager, and improve application startup and response speed.

Bug: 419835707

Change-Id: I73ad5b8d34a4c3f5d4e95cade33375368183eb83
Signed-off-by: zhanghao56 <zhanghao56@xiaomi.corp-partner.google.com>
Signed-off-by: Ghosuto <clash.raja10@gmail.com>